requirements.txt
pandas
numpy
scikit-learn
pyyaml
joblib
streamlit
gunicorn
-e .
